evict/requirements.txt
git+https://github.com/cloudwithax/pomice
git+https://github.com/MiroslavRosenov/better-ipc
git+https://github.com/x32u/uwuipy
git+https://github.com/x32u/Discord-iOS
git+https://github.com/parelite/discord.py
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
pydub
jishaku
PyNaCl
typing_extensions
humanize
rich
asyncpg
orjson
pydantic==1.9
humanfriendly
lru_dict
async_lru
redis
xxhash
beautifulsoup4
shazamio
python-dateutil
parsedatetime
aiohttp_cors
roblox
playwright
aiomisc
cryptography
msgpack
boltons
aiofile
tornado
psutil
wand
colorthief
httpx
asyncspotify
cashews
yt-dlp
pycryptodome
stackprinter
dateparser
arrow
xmltodict
cairosvg
asyncpraw
validators
rapidfuzz
lxml
colorama
aiomisc
pyparsing
pylast
wavelink
nudenet
gitpython
chat_exporter
matplotlib
posthog
scipy
boto3
stripe
plotly
moviepy
dnspython
unidecode